feature de delete agora remove também da web api. Funcionalidade. Adicionado também flags de desativado para essa feature funcionar mesmo quando o usuario estiver sem acesso a internet

+++ app/src/main/java/br/com/alura/ceep/repository/NotaRepository.kt
@@ -4,6 +4,7 @@ import br.com.alura.ceep.database.dao.NotaDao
 import br.com.alura.ceep.model.Nota
 import br.com.alura.ceep.webclient.NotaWebClient
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.first
 
 class NotaRepository(
     private val dao: NotaDao,
@@ -14,22 +15,43 @@ class NotaRepository(
         return dao.buscaTodas()
     }
 
-    suspend fun atualizaTodas() {
-        webClient.buscaTodas()?.let { listaNotas ->
-            dao.salva(listaNotas)
-        }
-    }
-
     fun buscaPorId(id: String): Flow<Nota> {
         return dao.buscaPorId(id)
     }
 
     suspend fun remove(id: String) {
-        dao.remove(id)
+        dao.desativa(id)
+        if (webClient.remove(id)) {
+            dao.remove(id)
+        }
     }
 
     suspend fun salva(nota: Nota) {
         dao.salva(nota)
-        webClient.salvar(nota)
+        if (webClient.salvar(nota)) {
+            val notaSincronizada = nota.copy(sincronizado = true)
+            dao.salva(notaSincronizada)
+        }
+    }
+
+    suspend fun sincroniza() {
+        val notasDesativadas = dao.buscaDesativadas().first()
+        notasDesativadas.forEach { notaDesativada ->
+            remove(notaDesativada.id)
+        }
+        val notasNaoSincronizadas = dao.buscaNotasNaoSincronizadas().first()
+        notasNaoSincronizadas.forEach { notaNaoSincronizada ->
+            salva(notaNaoSincronizada)
+        }
+        atualizaTodas()
+    }
+
+    private suspend fun atualizaTodas() {
+        webClient.buscaTodas()?.let { notas ->
+            val notasSincronizadas = notas.map { nota ->
+                nota.copy(sincronizado = true)
+            }
+            dao.salva(notasSincronizadas)
+        }
     }
 }
